+++ .eslintrc.cjs
@@ -3,7 +3,7 @@ const { builtinModules } = require('node:module');
 module.exports = {
 	extends: [
 		'plugin:@typescript-eslint/recommended-type-checked',
-		'plugin:@typescript-eslint/stylistic-type-checked',
+		'prettier',
 	],
 	parser: '@typescript-eslint/parser',
 	parserOptions: {
@@ -14,9 +14,10 @@ module.exports = {
 	rules: {
 		// These off/configured-differently-by-default rules fit well for us
 		'no-only-tests/no-only-tests': 'error',
-		'@typescript-eslint/no-shadow': ['error'],
+		'@typescript-eslint/no-shadow': 'off',
 		'no-console': 'off',
-
+		'@typescript-eslint/no-unused-vars': 'off',
+		'@typescript-eslint/ban-types': 'off',
 		// Todo: do we want these?
 		'@typescript-eslint/no-unused-vars': 'off',
 		'@typescript-eslint/ban-types': 'off',
@@ -63,12 +64,6 @@ module.exports = {
 			},
 			rules: {
 				'no-console': 'off',
-			},
-		},
-		{
-			files: ['packages/integrations/**/*.ts'],
-			rules: {
-				'no-console': ['error', { allow: ['warn', 'error', 'info', 'debug'] }],
 			},
 		},
 	],