+++ packages/cloudflare/test/fixtures/dev-runtime-pages/astro.config.ts
@@ -7,37 +7,45 @@ export default defineConfig({
 			mode: 'local',
 			type: 'pages',
 			bindings: {
-				"KV": {
-					type: "kv",
+				KV: {
+					type: 'kv',
 				},
-				"KV_PROD": {
-					type: "kv",
+				KV_PROD: {
+					type: 'kv',
 				},
-				"COOL": {
-					type: "var",
-					value: "ME"
+				COOL: {
+					type: 'var',
+					value: 'ME',
 				},
-				"D1": {
-					type: "d1",
+				D1: {
+					type: 'd1',
 				},
-				"D1_PROD": {
-					type: "d1",
+				D1_PROD: {
+					type: 'd1',
 				},
-				"R2": {
-					type: "r2",
+				R2: {
+					type: 'r2',
 				},
-				"R2_PROD": {
-					type: "r2",
+				R2_PROD: {
+					type: 'r2',
 				},
-				"DO": {
-					type: "durable-object",
-					className: "DO",
+				DO: {
+					type: 'durable-object',
+					className: 'DO',
 				},
-				"DO_PROD": {
-					type: "durable-object",
-					className: "DO_PROD",
+				DO_PROD: {
+					type: 'durable-object',
+					className: 'DO_PROD',
+				},
+				AUTH: {
+					type: 'service',
+					address: '127.0.0.1:8787'
+				},
+				LOGOUT: {
+					type: 'service',
+					address: '127.0.0.1:8787'
 				}
-			}
+			},
 		},
 	}),
 	output: 'server',