+++ packages/netlify/test/functions/redirects.test.js
@@ -1,5 +1,7 @@
+import { createServer } from 'http';
+import * as assert from 'node:assert/strict';
+import { before, describe, it } from 'node:test';
 import { loadFixture } from '@astrojs/test-utils';
-import { expect } from 'chai';
 
 describe('SSR - Redirects', () => {
 	let fixture;
@@ -12,8 +14,9 @@ describe('SSR - Redirects', () => {
 	it('Creates a redirects file', async () => {
 		const redirects = await fixture.readFile('./_redirects');
 		const parts = redirects.split(/\s+/);
-		expect(parts).to.deep.equal(['', '/other', '/', '301', '']);
-		expect(redirects).to.matchSnapshot();
+		assert.deepEqual(parts, ['', '/other', '/', '301', '']);
+		// Snapshots are not supported in Node.js test yet (https://github.com/nodejs/node/issues/48260)
+		assert.equal(redirects, '\n/other    /       301\n');
 	});
 
 	it('Does not create .html files', async () => {
@@ -23,6 +26,38 @@ describe('SSR - Redirects', () => {
 		} catch {
 			hasErrored = true;
 		}
-		expect(hasErrored).to.equal(true, 'this file should not exist');
+		assert.equal(hasErrored, true, 'this file should not exist');
+	});
+
+	it('renders static 404 page', async () => {
+		const entryURL = new URL(
+			'./fixtures/redirects/.netlify/functions-internal/ssr/ssr.mjs',
+			import.meta.url
+		);
+		const { default: handler } = await import(entryURL);
+		const resp = await handler(new Request('http://example.com/nonexistant-page'), {});
+		assert.equal(resp.status, 404);
+		assert.equal(resp.headers.get('content-type'), 'text/html; charset=utf-8');
+		const text = await resp.text();
+		assert.equal(text.includes('This is my static 404 page'), true);
+	});
+
+	it('does not pass through 404 request', async () => {
+		let testServerCalls = 0;
+		const testServer = createServer((req, res) => {
+			testServerCalls++;
+			res.writeHead(200);
+			res.end();
+		});
+		testServer.listen(5678);
+		const entryURL = new URL(
+			'./fixtures/redirects/.netlify/functions-internal/ssr/ssr.mjs',
+			import.meta.url
+		);
+		const { default: handler } = await import(entryURL);
+		const resp = await handler(new Request('http://localhost:5678/nonexistant-page'), {});
+		assert.equal(resp.status, 404);
+		assert.equal(testServerCalls, 0);
+		testServer.close();
 	});
 });